* Added deadlock detection for the fd table mutex * Correctly pass program defined LD_LIBRARY_PATH to the MVEE_LD_Loader (this broke firefox' plugin-container) * Correctly return -ENOENT when an execve of a non-existing file is attempted (this broke dbus-launch) * Fixed default path of the GHUMVEE-ready libstdc++ * Fixed a bug that caused some cloexec files to remain in the GHUMVEE fd table even after sys_execve * Fixed false positives in fd table verification * Changed syslock policy for sys_open/sys_openat (firefox has sys_open calls that block...) * Added reading envp vars from sys_execve calls * Added partial sys_unshare support * Added partial sys_seccomp support * Allow sys_recvmsg calls with msg_iovlen set to 0 * Silenced several compiler warnings

+++ MVEE/Src/MVEE.cpp
@@ -191,7 +191,7 @@ void mvee::init_aliases()
 		for (auto alias : variant_config["alias"])
 		{
 			auto str = alias.asString();
-			int pos = str.find("=");
+			size_t pos = str.find("=");
 			if (pos != std::string::npos)
 			{
 				std::string pattern     = str.substr(0, pos);
